SalaryGradeQuartilesBLL.cs: 	Added several new methods: 	 	public static SalaryGradeQuartiles GetRecord(string salaryGrade, DateTime effectiveDate): Get a single,
specific SalaryGradeQuartiles based on the salaryGrade and effectiveDate provided.
		public static SalaryGradeQuartiles GetRecord(SalaryGradeQuartiles example): Get the single 

SalaryGradeQuartiles, which matches the example provided.
		public static IList<SalaryGradeQuartiles> GetAll(string salaryGrade, string propertyName, bool ascending): 

Gets a list of all the SalaryGradeQuartiles matching the salaryGrade provided or all if no salaryGrade was provided, sorted 

by, and in the order provided.
		public static IList<SalaryGradeQuartiles> GetDistinct(): This method calls the underlying DAO method, which 

returns a distinct list of SalaryGradeQuartiles objects, one of each with the earliest EffectiveDate. This is because the 

list is built by taking the first object for each salary grade from the larger, complete list, which was sorted by 

EffectiveDate.  Restated, it returns a list of SalaryGradeQuartiles objects with distinct SalaryGrades, each having the 

earliest EffectiveDate out of all other SalaryGradeQuartiles  with the same Salary Grade.</returns>
		public static IList<String> GetDistinctSalaryGrades(): This method calls the underlying DAO method, which 

returns a distinct list of just the salary grades (strings) that can be used for populating a drop-down list with distinct 

salary grades. 
		
		
SalaryScaleBLL.cs:
	Revised method public static bool DeleteRecord(SalaryScale record, bool forceDelete) to delete any unique 

SalaryGradeQuartiles or just remove the SalaryScale from the SalaryGradeQuartiles.SalaryScales arrary if shared between 

SalaryScales.
	Added method public static IList<SalaryScale> GetSalaryScales(string titleCode, string propertyName, bool ascending), 

which given a title code, returns a list of SalaryScales, sorted by the propertyName in the order provided.  If the title 
 (Migrated by TFS Migration Toolkit) (Updated by TFS to TFS Connector, server: http://1x:8080/, id 2001. )

git-tfs-id: [https://tfs.caes.ucdavis.edu:8088/tfs/DefaultCollection]$/ESRA/Trunk/Esra;C5320