Unit: Added Users = new List<Users>();  to constructor.
EmployeeSalaryComparison controller: Added return result values for all items that change on the server-side so that they might be updated on the client-side after the Ajax save.
EmployeeSalaryComparison/Details: Added logic to use associative array and data field name to decode data table cell number instead of cell number itself.  That way the position can change and the cell number not need to be updated.