Worked NHibernate profiler and made modifications to limit number of database accesses to 1 for each type of objects, i.e. Employees, Departments, Titles, Etc.
This was mainly achieved by fetching all objects of a particular type and then using LINQ to do any additional filtering and sorting, plus using any previously fetched lists in subsequent methods.

Moved some of the "fetching" logic from the controller to the model in order to minimize amount of code duplication.

Got most of the LINQ many-to-many queries working properly.